requirements.txt
streamlit>=1.38.0
boto3>=1.35.0
pandas>=2.2.0
python-dotenv>=1.0.1
